Updated support for -n -u --force
As of Hashcat v3.00 direct use of -n and -u was [no longer supported ](https://hashcat.net/forum/thread-5551-post-29892.html#pid29892) however, while the commands -n -u are still recognized, depending on the release version, Hashcat no longer recognize the inputs for `--gpu-loops` and `--gpu-accel` but will recognize -n and -u following a notice that the user may use '--force' to proceed with hang-tuned settings anyway. Though it's a bit sloppy (due to lack of any formal training/education), I've made changes so that if the user adds -n or -u it will add the options as -n and -u respectively, instead of the long option. I've also added a short option -R as part of the --force option with a warning to the user not to report errors. I did not, however, create a conditional statement in  case the user fails to add the `--force` option when adding the -n and/or -u options nor will it check for the absence of the `-w --workload-profile` before adding said commands simply because it was not present before. Not sure if it's worth adding but I found these extremely useful when having allowed a hash to work for a period of time before discovering the commands and not wanting to have to start over. Thank you so much for making this!